Moved models into ./models/ directory Removed temp.rb
