+++ test/models/model_with_enc.rb
