+++ test/models/my_model.rb
@@ -1,4 +1,4 @@
-require File.expand_path(File.dirname(__FILE__) + "/../lib/simple_record")
+require File.expand_path(File.dirname(__FILE__) + "/../../lib/simple_record")
 require_relative 'my_base_model'
 require_relative 'my_sharded_model'
 require 'active_model'
