+++ test/models/my_sharded_model.rb
@@ -1,6 +1,4 @@
-require File.expand_path(File.dirname(__FILE__) + "/../lib/simple_record")
-
-require File.expand_path(File.dirname(__FILE__) + "/../lib/simple_record")
+require File.expand_path(File.dirname(__FILE__) + "/../../lib/simple_record")
 
 class MyShardedModel < SimpleRecord::Base
