+++ test/test_base.rb
@@ -4,8 +4,8 @@ require File.join(File.dirname(__FILE__), "/../lib/simple_record")
 require File.join(File.dirname(__FILE__), "./test_helpers")
 require "yaml"
 require 'aws'
-require_relative 'my_model'
-require_relative 'my_child_model'
+require_relative 'models/my_model'
+require_relative 'models/my_child_model'
 #require 'active_support'
 
 class TestBase < Test::Unit::TestCase
