+++ test/test_json.rb
@@ -4,9 +4,9 @@ require File.join(File.dirname(__FILE__), "./test_helpers")
 require File.join(File.dirname(__FILE__), "./test_base")
 require "yaml"
 require 'aws'
-require_relative 'my_model'
-require_relative 'my_child_model'
-require_relative 'model_with_enc'
+require_relative 'models/my_model'
+require_relative 'models/my_child_model'
+require_relative 'models/model_with_enc'
 require 'active_support/core_ext'
 
 # Tests for SimpleRecord
