+++ test/test_lobs.rb
@@ -4,9 +4,9 @@ require_relative "test_helpers"
 require_relative "test_base"
 require "yaml"
 require 'aws'
-require_relative 'my_model'
-require_relative 'my_child_model'
-require_relative 'model_with_enc'
+require_relative 'models/my_model'
+require_relative 'models/my_child_model'
+require_relative 'models/model_with_enc'
 
 # Tests for SimpleRecord
 #
