+++ test/test_marshalled.rb
@@ -2,8 +2,8 @@ require 'test/unit'
 require File.expand_path(File.dirname(__FILE__) + "/../lib/simple_record")
 require "yaml"
 require 'aws'
-require 'my_model'
-require 'my_child_model'
+require_relative 'models/my_model'
+require_relative 'models/my_child_model'
 require 'active_support'
 require 'test_base'
